fable: Simplify make_prototype definition file
- All outlined definitions that use make_prototype are inlined
  again in their respective header files.

BREAKING CHANGE: The function `make_const_str` is removed in favor of
`make_const_schema`. This is orthogonal with `make_schema` and works
the same as `make_const_str`, except that you need to make sure a
`std::string` is actually used instead of something else, such as
`char *` or `std::string_view`.

For example, the following statement:

    make_const_str("constant");

should become:

    use namespace std::literals;
    make_const_schema("constant"s);

or:

    make_const_schema(std::string("constant"));
