+++ fable/include/fable/schema/optional.hpp
@@ -72,17 +72,13 @@ class Optional : public Base<Optional<T, P>> {
   using ValueType = typename Type::value_type;
   using PrototypeSchema = std::remove_cv_t<std::remove_reference_t<P>>;
 
-  Optional(Type* ptr, std::string desc);
+  Optional(Type* ptr, std::string desc)
+      : Optional(ptr, make_prototype<typename T::value_type>(), std::move(desc)) {}
+
   Optional(Type* ptr, PrototypeSchema prototype, std::string desc)
       : Base<Optional<T, P>>(prototype.type(), std::move(desc)), prototype_(std::move(prototype)), ptr_(ptr) {
     prototype_.reset_ptr();
   }
-
-#if 0
-  // This is defined in: fable/schema/xmagic.hpp
-  Optional(T* ptr, std::string desc)
-    : Optional<T, P>(ptr, make_prototype<typename T::value_type>(), std::move(desc)) {}
-#endif
 
  public:  // Overrides
   std::string type_string() const override { return prototype_.type_string() + "?"; }
@@ -151,6 +147,11 @@ class Optional : public Base<Optional<T, P>> {
 template <typename T, typename P, typename S, std::enable_if_t<is_optional_v<T>, bool> = true>
 inline Optional<T, P> make_schema(T* ptr, P&& prototype, S&& desc) {
   return Optional<T, P>(ptr, std::forward<P>(prototype), std::forward<S>(desc));
+}
+
+template <typename T, typename S, std::enable_if_t<is_optional_v<T>, bool> = true>
+Optional<T, decltype(make_prototype<typename T::value_type>())> make_schema(T* ptr, S&& desc) {
+  return Optional<T, decltype(make_prototype<typename T::value_type>())>(ptr, std::forward<S>(desc));
 }
 
 }  // namespace schema
