+++ fable/src/fable/schema_test.cpp
@@ -114,6 +114,7 @@ struct MyStruct : public fable::Confable {
 TEST(fable_schema, schema_wrapper) {
   using namespace fable;          // NOLINT(build/namespaces)
   using namespace fable::schema;  // NOLINT(build/namespaces)
+  using namespace std::literals;  // NOLINT(build/namespaces)
 
   bool my_required = false;
   std::string my_string = "";
@@ -126,7 +127,7 @@ TEST(fable_schema, schema_wrapper) {
   std::optional<std::string> middlename;
 
   auto s1 = Schema{
-      {"author", make_const_str("me", "author of this code")},
+      {"author", make_const_schema("me"s, "author of this code")},
       {"required", make_schema(&my_required, "my required boolean, should be true").require()},
       {"string", Schema(&my_string, "my string")},
       {"int", make_schema(&my_int, "my integer").minimum(0)},
