+++ runtime/include/cloe/core/fable.hpp
@@ -36,7 +36,6 @@ using fable::Schema;
 using fable::SchemaError;
 
 using fable::make_const_schema;
-using fable::make_const_str;
 using fable::make_prototype;
 using fable::make_schema;
 using fable::schema_type;
